shop: 8.4.3   * Auto-replacement of a national direct-dialing (NDD) code with a country code,     specified in site login and signup settings, in customers’ phone numbers.   * Unified identification of phone numbers with an NDD code and a country code     during in-cart checkout and in backend’s “Customers” section.   * Support for in-cart checkout in products’ “Buy” widgets.   * Ignoring of whitespace characters in customers’ email addresses during checkout.   * Fixed deletion of customers’ shipping addresses during multi-step checkout     with enabled option “Add an order to existing customer profile with the same email address”.   * Improved editing and creation of orders in the backend:     * Unified identification of phone numbers with an NDD code and a country code       specified in login and signup settings of the order storefront’s website.     * Default selection of the country from general settings in the payment address form.     * Fixed displaying of available fields for “New customer” mode.     * Availability of “Manual order” source option during order editing.     * Additional shipping address fields hidden.     * “Title”, “Job title”, and “Company” fields hidden for “Person” type customers.       “Job title” field hidden for “Company” type customers.   * Improved displaying of customer profile fields in “Customers” section.   * Fixed displaying of company icons and Gravatar userpics.   * Fixed functioning of {$wa->shop->settings()} method in Site app’s design themes     with in-cart checkout selected in Shop-Script storefront settings.   * Improved “Quick start” section with new options to load demo products and settings.   * “Products” section improvements:     * New sorting options by main SKU’s code, compare at and purchase prices.     * Displaying of product SKU IDs for convenient import of CSV files.     * Displaying of SKU codes under “Services” tab on product-editing page.     * Improved creation of product duplicates.     * Improved user interface of SKU deletion on product-editing page.     * Fixed displaying of color markers for different stock levels in “SKUs” view mode.     * Fixed product category filter settings errors.     * Fixed status changing for all products displayed via a tag cloud link.     * Fixed search of “Multiple values (checkboxes)” type product features in column settings       for “Table” view mode.     * Fixed minor error in product set creation user interface.   * CSV file export & import:     * Selection of exported product image size.     * Export and import of product images and SKUs association. Upon export, be sure to add       an extra line for main product SKUs with their features and images.     * Export and import of product image descriptions.     * Setting to manage import of product images by URLs containing equal file names or paths.     * Fixed import of products with few columns in a CSV file.     * Fixed “Invalid category count” error during export.     * Fixed product export from dynamic categories.     * Fixed product identification by ID.     * Fixed excessive export of SKUs’ weight values with their product weight.     * Fixed duplication of path to Webasyst in product image URLs when Webasyst is installed       in a website subdirectory.

+++ lib/classes/shopProduct.class.php
@@ -1187,6 +1187,7 @@ class shopProduct implements ArrayAccess
         $skip = array(
             'id',
             'create_datetime',
+            'edit_datetime',
             'id_1c',
             'rating',
             'rating_count',
@@ -1244,6 +1245,10 @@ class shopProduct implements ArrayAccess
                     $storage_data = array();
                     $i = 0;
                     foreach ($raw as $sku_id => $sku) {
+                        if (isset($sku['sku']) && ifset($options, 'remove_sku', true)) {
+                            $sku['sku'] = null;
+                        }
+
                         if (!empty($sku['virtual']) || $ignore_select) {
                             if ($file_path = shopProductSkusModel::getPath($sku)) {
                                 $sku_files[$sku['id']] = array(
@@ -1282,7 +1287,10 @@ class shopProduct implements ArrayAccess
 
         $counter = 0;
         $data['url'] = shopHelper::genUniqueUrl($this->url, $this->model, $counter);
-        $data['name'] = $this->name.sprintf(' %d', $counter ? $counter : 1);
+
+        if (ifset($options, 'change_name', true)) {
+            $data['name'] = $this->name.sprintf(' %d', $counter ? $counter : 1);
+        }
 
         if (!$duplicate->save($data, true, $errors)) {
             return false;
@@ -1389,8 +1397,10 @@ class shopProduct implements ArrayAccess
             }
         }
 
-        // Delete the empty values. Values are stored in the $duplicate->save
-        if ($delete_features_id) {
+        // Delete the empty values. Reset only if sku are created in "Parameter selection"
+        // For this mode, the characteristics included for the product break the filtering
+        // Values are stored in the $duplicate->save
+        if ($delete_features_id && $data['sku_type'] == 1) {
             $product_features_model->deleteByField(array('product_id' => $product_id, 'feature_id' => $delete_features_id));
         }
